Feature/use settings object (#1664)
* use settings object

* more services use settings.go

* more services use settings.go

* lint fix

* yet more lint fixes

* fix failing test

* use utils getUrl for utxo store setting

* refactor block validation server to use the settings object in the right way

* lint fix

* small changes

* more services use settings.go

* linter fixes

* debug for failing smoke test

* use different settings configs for nodes

* fix smoke test

* final services use settings object

* more refactor to blockvalidation server, moved all of Init() into New()

* Revert "more refactor to blockvalidation server, moved all of Init() into New()"

This reverts commit 0a1d8584743c569fe0fc826cb7de30f6241267d7.

* fix e2e test by allowing min coinbase fee as 0

* more small changes

* use settings in some cmd files

* move blockValidation init back

* fix test

* fix smoke test setup

---------

+++ services/validator/data.go
@@ -79,7 +79,7 @@ func (d *TxValidationData) Bytes() []byte {
 
 	// Write height (4 bytes) in little-endian format
 	b32 := make([]byte, 4)
-	binary.LittleEndian.PutUint32(b32, uint32(d.Height))
+	binary.LittleEndian.PutUint32(b32, d.Height)
 	bytes = append(bytes, b32...)
 
 	// Append transaction data
